Download Manager Improvements
- Added continue after failure, pause after failure, and pause after N failures
    - Also added parameters to control/choose behavior
- Fixed unresponsiveness after a worker fails during a parallel download
- Added around 90 more tests
- Added URL validation and updated output_file name validation
- Better download speed calculation for GUI
- More reliable shutdown
- Fixed other minor bugs and added minor improvements to download manager
    - Possibility of async tasks going untracked
    - Allow input number of workers on adding download
    - Simplify increment for parallel download segments
    - Removed data_queues and use a range_iterator for download segments
    - Separate Parallel Download Metadata from Download Metadata so single downloads don't have to initialize PD Metadata specific structures
    - And likely more...
- Restructured the repo to be the Download Manager repo and not the Practice repo
- Updated CI to use Windows environment NOT Ubuntu

+++ tests/conftest.py
@@ -3,18 +3,22 @@ import pytest
 import os
 import logging
 import copy
+import time
 
 from typing import Dict
-from dmanager.asyncio_thread import AsyncioEventLoopThread
 
+from dmanager.asyncio_thread import AsyncioEventLoopThread
+from dmanager.core import DownloadManager
+from dmanager.constants import SEGMENT_SIZE
 
 class MockResponse:
-    def __init__(self, status, headers={}):
+    def __init__(self, status, headers=None):
         self.status = status
         self.headers = headers
         self.content = self
         self.queue = asyncio.Queue()
         self.stop = False
+        self.exception = None
 
     async def __aenter__(self):
         return self
@@ -24,6 +28,9 @@ class MockResponse:
 
     async def iter_chunked(self, chunk_size_limit):
         while not self.stop or not self.queue.empty():
+            if self.exception is not None:
+                raise self.exception
+
             if not self.queue.empty():
                 chunk = await self.queue.get()
                 yield chunk
@@ -40,10 +47,14 @@ class MockResponse:
         while not self.queue.empty():
             await self.queue.get()
 
+    def set_exception(self, exception:Exception):
+        self.exception = exception
+
 
 class MockSession:
     def __init__(self, responses):
         self._responses = responses
+        self.closed = False
 
     async def __aenter__(self):
         return self
@@ -58,6 +69,7 @@ class MockSession:
         return self._responses[url]
     
     async def close(self):
+        self.closed = True
         return
 
 
@@ -68,8 +80,9 @@ class MockParallelResponse():
         self.content = self
         self.request_queue = request_queue
         self.data = copy.deepcopy(data)
-        self.send_next_letter: Dict[str: int] = {x:0 for x in range_ends}
+        self.send_n_letters: Dict[str, int] = {x:0 for x in range_ends}
         self.done = {x: False for x in range_ends}
+        self.exception = None
 
     async def __aenter__(self):
         return self
@@ -82,25 +95,31 @@ class MockParallelResponse():
             return
         data_range_request = await self.request_queue.get()
         range_end = data_range_request.split("-")[-1]
-        while True:
-            if range_end not in self.send_next_letter:
-                    raise Exception(f"Got unexpected {range_end=}")
-            if self.send_next_letter[range_end] > 0:
+
+        while True:            
+            if self.exception is not None:
+                raise self.exception
+
+            if range_end not in self.send_n_letters:
+                raise Exception(f"Got unexpected {range_end=}")
+            
+
+            if self.send_n_letters[range_end] > 0:
                 if len(self.data[range_end]) > 0:
-                    slice_size = min(1024*256, len(self.data[range_end]))
+                    slice_size = min(1024, self.send_n_letters[range_end])
                     yield bytes(self.data[range_end][:slice_size])
                     self.data[range_end] = self.data[range_end][slice_size:]
-                    self.send_next_letter[range_end] -= slice_size
+                    self.send_n_letters[range_end] -= slice_size
                 else:
                     # No more data to send
                     break
             else:
                 await asyncio.sleep(1)
-            if self.done[range_end] and self.send_next_letter[range_end] == 0:
+            if self.done[range_end] and self.send_n_letters[range_end] == 0:
                 break
     
     def set_range_end_n_send(self, range_end, n):
-        self.send_next_letter[range_end] = n
+        self.send_n_letters[range_end] = n
     
     def set_range_end_done(self, range_end):
         self.done[range_end] = True
@@ -108,15 +127,14 @@ class MockParallelResponse():
     def end_response(self):
         self.stop = True
 
-    async def empty_queue(self):
-        while not self.queue.empty():
-            await self.queue.get()
-
+    def set_exception(self, exception:Exception):
+        self.exception = exception
 
 class MockParallelSession:
     def __init__(self, responses, request_queues):
         self._responses = responses
         self.request_queues = request_queues
+        self.closed = False
 
     async def __aenter__(self):
         return self
@@ -132,14 +150,19 @@ class MockParallelSession:
         return self._responses[url]
 
     async def close(self):
+        self.closed = True
         return
 
-
 @pytest.fixture
-def async_thread_runner():
+def async_thread_runner(request):
     runner = AsyncioEventLoopThread()
-    yield runner
-    runner.shutdown()
+    
+    def cleanup():
+        logging.debug("Async thread fixture shutting down.")
+        runner.shutdown()
+
+    request.addfinalizer(cleanup)
+    return runner
 
 @pytest.fixture
 def create_mock_response_and_set_mock_session(monkeypatch):
@@ -192,7 +215,7 @@ def test_file_setup_and_cleanup(request):
             os.remove(test_file_name)
 
     def cleanup():
-        if os.path.exists(test_file_name):
+        if test_file_name != "" and os.path.exists(test_file_name):
             logging.debug(f"Cleaning up {test_file_name=}")
             os.remove(test_file_name)
     
@@ -216,7 +239,49 @@ def test_multiple_file_setup_and_cleanup(request):
             if os.path.exists(file):
                 logging.debug(f"Cleaning up {file=}")
                 os.remove(file)
+                
     
     request.addfinalizer(cleanup)
     yield setup
 
+@pytest.fixture
+def download_manager_fixture(request):
+    download_manager = None
+    runner = None
+
+    def setup(
+        async_runner: AsyncioEventLoopThread,
+        running_event_update_rate_seconds: int = 1, 
+        parallel_running_event_update_rate_seconds: int = 1, 
+        maximum_workers_per_task: int = 5, 
+        request_connect_timeout: float= 120.,
+        request_read_timeout: float= 60.,
+        request_total_timeout: float = None,
+        parallel_download_segment_size: int= SEGMENT_SIZE,
+        continue_on_error: bool= True,
+        stop_continue_on_n_errors: int | None=5
+    ):
+        nonlocal runner
+        nonlocal download_manager
+        runner = async_runner
+        download_manager = DownloadManager(
+            running_event_update_rate_seconds=running_event_update_rate_seconds,
+            parallel_running_event_update_rate_seconds=parallel_running_event_update_rate_seconds,
+            maximum_workers_per_task=maximum_workers_per_task,
+            request_connect_timeout=request_connect_timeout,
+            request_read_timeout=request_read_timeout,
+            request_total_timeout=request_total_timeout,
+            parallel_download_segment_size=parallel_download_segment_size,
+            continue_on_error=continue_on_error,
+            stop_continue_on_n_errors=stop_continue_on_n_errors
+        )
+        return download_manager
+
+    def cleanup():
+        logging.debug("Download Manager fixture cleanup.")
+        if download_manager:
+            future = runner.submit(download_manager.shutdown())
+            future.result(timeout=15)
+    
+    request.addfinalizer(cleanup)
+    yield setup
